requirements.txt
appdirs==1.4.4
asgiref==3.8.1
baron==0.10.1
dj-database-url==2.3.0
dj-email-url==1.0.6
Django==5.1.4
django-autoslug==1.9.9
django-cache-url==3.4.5
django-environs==0.1.1
django-model-utils==5.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
environs==11.2.1
gunicorn==23.0.0
inflection==0.5.1
marshmallow==3.23.2
packaging==24.2
pillow==11.0.0
PyJWT==2.10.1
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redbaron==0.9.2
rply==0.7.8
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
